unit3

+++ APCSA/src/Unit2/Line.java
@@ -5,10 +5,10 @@
 package Unit2;
 public class Line
 {
-	public static double getSlope( double x1, double y1, double x2, double y2 )
+	public static double getSlope( int x1, int y1, int x2, int y2 )
 	{
 		
-		return (y2-y1)/(x2-x1);
+		return ((double)y2-(double)y1)/((double)x2-(double)x1);
 		//return 0;
 	}
